refactor: convention-default helpers live in base.py, not yaml_source
OssieSource had been importing _parse_convention_default and
_apply_convention_default -- two private functions -- across from
yaml_source.py. Every other helper shared between the sources lives in
base.py and is imported by both: validate_decompositions,
validate_drill_by, parse_review_date, jsonify_extras,
build_relationship_index.

Two things made the old placement the odd one out. Both helpers operate
purely on MetricDefinition / Decomposition and touch nothing
YAML-specific, and both had to import VALID_CONVENTIONS and
_CROSS_TERM_OPERATORS back out of base.py to do their work. After the
move yaml_source.py needs neither constant for anything else -- it had no
other consumer of them -- which is the clearest evidence the vocabulary
they enforce already lived where the functions belonged.

parse_review_date's docstring states the principle the move follows: it
is shared rather than per-source because every format reaching the same
field reaches the same arithmetic, "so a second copy of this would be a
second chance to get the subclass ordering wrong." Both sources resolve a
house default onto the same Decomposition.convention field, and
fold_into-is-not-a-valid-default is exactly the rule that drifts if it is
ever copied.

No behavior change. Both source paths keep their existing tests, and
verified directly that a YAML default still resolves (split_evenly), an
Ossie vendor-block default still resolves (explicit), a fold_into
source-level default is still rejected, and the pinned round-trip digest
is unmoved. 1078 passing.

No version bump: both functions are private, so nothing on the public
surface changed. Rides along with the 0.43.1 already on main.

Closes #74
