+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 
 ### Internal
 
+- **The convention-default helpers moved to `semantic/base.py`.** `OssieSource` had been importing `_parse_convention_default` / `_apply_convention_default` — two *private* functions — across from `yaml_source.py`, while every other helper shared between sources (`validate_decompositions`, `validate_drill_by`, `parse_review_date`, `jsonify_extras`, `build_relationship_index`) lives in `base.py`. The giveaway was that both had to import `VALID_CONVENTIONS` and `_CROSS_TERM_OPERATORS` back out of `base.py` to do their work, and that those two constants had no other consumer in `yaml_source.py` — the vocabulary they enforce already lived where the functions belonged. `parse_review_date`'s own docstring states the principle: a second copy is a second chance to get the rule wrong, and both sources resolve a default onto the same `Decomposition.convention` field. No behavior change; both source paths keep their existing tests. ([#74](https://github.com/flyersworder/agentic-data-contracts/issues/74))
 - **The pinned round-trip digest now covers metric serialization.** `_PINNED_ROUNDTRIP_DIGEST` is the repo's broadest guard against canonical-bytes drift, but it was computed over a semantic source with no `metrics:` key at all, so `_dump_metric` was never called and it could not see any change to metric serialization. That is worse than a missing test, because the assertion passed either way: every optional field added to `_dump_metric` since the pin was written could have been made unconditional — moving the digest of every real contract and invalidating every published ARD attestation — with the test still green. The fixture now carries a metric exercising every optional field the dump writes plus two leaves for the omit-when-empty branches, verified by mutation (removing the `if m.decompositions:` guard now fails the pin), and a companion test asserts the pinned contract's canonical bytes still carry those fields so the coverage cannot be silently narrowed again. No shipped code changed. ([#73](https://github.com/flyersworder/agentic-data-contracts/issues/73))
 
 ## [0.43.0] - 2026-08-18
